return 12586269025L given 50, and use long type to replace int.

+++ src/main/java/cc/xpbootcamp/warmup/fibonacci/Fibonacci.java
@@ -2,10 +2,10 @@ package cc.xpbootcamp.warmup.fibonacci;
 
 public class Fibonacci {
 
-  public static int getFibonacciByDigits(int digits) {
+  public static long getFibonacciByDigits(int digits) {
 
     if (digits == 1 || digits == 2) {
-      return 1;
+      return 1L;
     }
 
     return getFibonacciByDigits(digits - 1) + getFibonacciByDigits(digits - 2);
